requirements.txt
mysqlclient==2.1.1
SQLAlchemy==1.4.41
